FaceDetector/app/requirements.txt
Flask
opencv-python
Pillow
requests
numpy